requirements.txt
Django==4.2.1
django-crispy-forms==2.0
django-bootstrap4==22.3
crispy-bootstrap4==2022.1
djangorestframework==3.14.0
mysqlclient